MSVC: Split point into separate x and y assignments
This avoids a compile error where you can't perform a narrowing conversion
in a C++11 initializer list, so we do the same thing without an initializer
list.  GCC doesn't mind either way.

+++ src/fmt-map-nukem2.cpp
@@ -871,7 +871,7 @@ class Map_Nukem2: public MapCore, public Map2DCore
 
 		virtual Point mapSize() const
 		{
-			return {this->mapWidth, 32750 / this->mapWidth};
+			return {(long)this->mapWidth, (long)(32750 / this->mapWidth)};
 		}
 
 		virtual void mapSize(const Point& newSize)
